# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_TrinityKnights_Backend_internal_domain_model.CreateVenueRequest:
    properties:
      address:
        maxLength: 255
        type: string
      capacity:
        type: integer
      city:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        type: string
      state:
        maxLength: 100
        type: string
      zip:
        maxLength: 10
        type: string
    required:
    - address
    - capacity
    - city
    - name
    - state
    - zip
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.LoginRequest:
    properties:
      email:
        maxLength: 100
        type: string
      password:
        maxLength: 255
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.PageMetadata:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.RegisterRequest:
    properties:
      email:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        type: string
      password:
        maxLength: 255
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  ? github_com_TrinityKnights_Backend_internal_domain_model.Response-array_github_com_TrinityKnights_Backend_internal_domain_model_VenueResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.VenueResponse'
        type: array
      error:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      paging:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.PageMetadata'
    type: object
  ? github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_TokenResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.TokenResponse'
      error:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      paging:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.PageMetadata'
    type: object
  ? github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_UserResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.UserResponse'
      error:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      paging:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.PageMetadata'
    type: object
  ? github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_VenueResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.VenueResponse'
      error:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      paging:
        $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.PageMetadata'
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.UpdateRequest:
    properties:
      email:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        type: string
      password:
        maxLength: 255
        minLength: 8
        type: string
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.UpdateVenueRequest:
    properties:
      address:
        maxLength: 255
        type: string
      capacity:
        type: integer
      city:
        maxLength: 100
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        type: string
      state:
        maxLength: 100
        type: string
      zip:
        maxLength: 10
        type: string
    required:
    - address
    - capacity
    - city
    - id
    - name
    - state
    - zip
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      status:
        type: boolean
      updated_at:
        type: string
    type: object
  github_com_TrinityKnights_Backend_internal_domain_model.VenueResponse:
    properties:
      address:
        type: string
      capacity:
        type: integer
      city:
        type: string
      id:
        type: integer
      name:
        type: string
      state:
        type: string
      zip:
        type: string
    type: object
info:
  contact:
    email: jakueenak@gmail.com
    name: API Support
  description: This is an auto-generated API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Trinity Knights API
  version: "0.1"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Get user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      security:
      - ApiKeyAuth: []
      summary: Get user profile
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      summary: Register a new user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update user profile
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - user
  /users/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      summary: Login user
      tags:
      - user
  /users/refresh:
    post:
      consumes:
      - application/json
      description: Refresh token
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      summary: Refresh token
      tags:
      - user
  /venues:
    get:
      description: Get a paginated list of all venues
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Sort field
        enum:
        - name
        - capacity
        in: query
        name: sort
        type: string
      - description: Sort order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-array_github_com_TrinityKnights_Backend_internal_domain_model_VenueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      security:
      - ApiKeyAuth: []
      summary: Get all venues
      tags:
      - venues
    post:
      consumes:
      - application/json
      description: Create a new venue with the provided details
      parameters:
      - description: Venue details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.CreateVenueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_VenueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      security:
      - ApiKeyAuth: []
      summary: Create a new venue
      tags:
      - venues
  /venues/{id}:
    get:
      description: Get details of a specific venue by its ID
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_VenueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      security:
      - ApiKeyAuth: []
      summary: Get a venue by ID
      tags:
      - venues
    put:
      consumes:
      - application/json
      description: Update an existing venue with the provided details
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated venue details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.UpdateVenueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-github_com_TrinityKnights_Backend_internal_domain_model_VenueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      security:
      - ApiKeyAuth: []
      summary: Update an existing venue
      tags:
      - venues
  /venues/search:
    get:
      description: Search venues with the provided query parameters
      parameters:
      - description: Venue name
        in: query
        name: name
        type: string
      - description: Venue address
        in: query
        name: address
        type: string
      - description: Venue capacity
        in: query
        name: capacity
        type: integer
      - description: Venue city
        in: query
        name: city
        type: string
      - description: Venue state
        in: query
        name: state
        type: string
      - description: Venue zip
        in: query
        name: zip
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Sort field
        enum:
        - name
        - capacity
        in: query
        name: sort
        type: string
      - description: Sort order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Response-array_github_com_TrinityKnights_Backend_internal_domain_model_VenueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_TrinityKnights_Backend_internal_domain_model.Error'
      security:
      - ApiKeyAuth: []
      summary: Search venues
      tags:
      - venues
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"